Corrected identification of valid codes and patients
Codes are now only used if they are to be kept and not ignored and they
are contained in the history of a patient that is to be kept and not
ignored. Similarly for patients, except that their history must contain a
code that is to be kept and not ignored.